// mobile/src/api/generated/schemas/planDayExerciseResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * GrowIt.API | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PlanDayExerciseResponse {
  id: string;
  /** @nullable */
  exerciseId: string | null;
  exerciseName: string;
  /** @nullable */
  category: string | null;
  targetSets: number;
  targetReps: number;
  restSeconds: number;
  orderIndex: number;
}