// mobile/src/api/generated/schemas/planDayExerciseSelection.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * GrowIt.API | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PlanDayExerciseSelection {
  /** @nullable */
  planDayExerciseId: string | null;
  /** @nullable */
  exerciseId: string | null;
  /** @nullable */
  customExerciseName: string | null;
}